+++ Encore/src/menus/ChartLoadingMenu.h
@@ -10,6 +10,8 @@ class ChartLoadingMenu : public OvershellMenu {
 public:
     ChartLoadingMenu() {};
     ~ChartLoadingMenu() {};
+    void KeyboardInputCallback(int key, int scancode, int action, int mods) override {};
+    void ControllerInputCallback(int joypadID, GLFWgamepadstate state) override {};
     void Draw() override;
     void Load() override;
 };
