+++ Encore/src/menus/ReadyUpMenu.h
@@ -10,6 +10,8 @@ class ReadyUpMenu : public OvershellMenu {
 public:
     ReadyUpMenu() = default;
     ~ReadyUpMenu() override = default;
+    void KeyboardInputCallback(int key, int scancode, int action, int mods) override;
+    void ControllerInputCallback(int joypadID, GLFWgamepadstate state) override;
     void Draw() override;
     void Load() override;
 };
